Use base handler for password reset job
